code cleanup
* The class name 'SpawnOptionEntity_1_13' doesn't match '[A-Z][a-zA-Z0-9]*'
* Replaced some fields by local variables
* Use one line for each declaration, it enhances code readability.
